pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpi-mcp4922"
version = "1.0.0"
description = "Python Library for driving MCP4922 DAC on Raspberry Pi"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "mrwunderbar666"}
]
keywords = ["raspberry-pi", "mcp4922", "dac", "spi", "gpio", "digital-to-analog"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Hardware",
    "Topic :: System :: Hardware :: Hardware Drivers",
]
requires-python = ">=3.7"
dependencies = [
    "RPi.GPIO>=0.7.0",
    "spidev>=3.4",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/mrwunderbar666/Python-RPi-MCP4922"
Repository = "https://github.com/mrwunderbar666/Python-RPi-MCP4922"
Issues = "https://github.com/mrwunderbar666/Python-RPi-MCP4922/issues"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PT"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 88
indent-width = 4

target-version = "py37"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = true
